+++ packages/sn-filter-pane/src/components/FoldedListbox/FoldedListbox.tsx
@@ -94,6 +94,7 @@ export const FoldedListbox = ({
     constraints, styles, options, translator, containerSize,
   } = stores.store.getState();
   const isRtl = options.direction === 'rtl';
+  const { qDimensionInfo } = resource.layout.qListObject;
   const {
     gridHeight, narrowSmall, narrowLarge,
   } = getSizes(containerSize, isInPopover);
@@ -142,7 +143,7 @@ export const FoldedListbox = ({
         {!narrowLarge
           && <Grid item width='100%' height={narrowSmall ? '100%' : undefined}>
             <SelectionSegmentsIndicator
-              qDimensionInfo={resource.layout.qListObject.qDimensionInfo} fullHeight={narrowSmall}
+              qDimensionInfo={qDimensionInfo} fullHeight={narrowSmall} styles={styles}
             ></SelectionSegmentsIndicator>
           </Grid>
         }