+++ packages/sn-filter-pane/src/hooks/use-styling/index.js
@@ -1,0 +1,3 @@
+import useStyling from './use-styling';
+
+export default useStyling;